+++ framework/model/src/main/java/is/codion/framework/model/EntityTableModel.java
@@ -21,7 +21,7 @@ package is.codion.framework.model;
 import is.codion.common.Configuration;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.model.FilterModel;
-import is.codion.common.model.selection.SelectionModel;
+import is.codion.common.model.selection.Selection;
 import is.codion.common.property.PropertyValue;
 import is.codion.common.state.State;
 import is.codion.common.value.Value;
@@ -202,9 +202,9 @@ public interface EntityTableModel<E extends EntityEditModel> extends FilterModel
 	int rowCount();
 
 	/**
-	 * @return the {@link SelectionModel}
+	 * @return the {@link Selection} instance
 	 */
-	SelectionModel<Entity> selection();
+	Selection<Entity> selection();
 
 	/**
 	 * @return the underlying query model